Rearranged the menu, should be much better now

+++ GLFW3/Inventory.cpp
@@ -132,7 +132,7 @@ void Inventory::display(Graphics::Text* text){
 }
 
 int Inventory::size(){
-    return held.size();
+    return (int)held.size();
 }
 
 int Inventory::getSelectedPos(){
